fix(outbound): fix clientpolicy not work after remvoe service profile
This is a preview PR for https://github.com/linkerd/linkerd2/issues/14274.
In linkerd/idle-cache/src/lib.rs get_or_insert_with function, seems the
sidercar with service profile and sidercar with client policy has the
same key, which cause the cache didn't expire after we delete
serviceprofile and create the new httproute for the target sevice for
pods that keep requesting the target service. this pr try to create
different keys for service profile httpsidecar and client policy httpsidecar